backend/requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
langgraph>=0.2.55
langgraph-checkpoint-sqlite>=2.0.0
langchain-groq>=0.2.1
langchain-core>=0.3.20
google-auth>=2.36.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.154.0
apscheduler>=3.10.4
sqlalchemy>=2.0.36
pdfplumber>=0.11.0
python-docx>=1.1.2
python-multipart>=0.0.12
python-dotenv>=1.0.1
pydantic>=2.10.3
